+++ commodore/component/__init__.py
@@ -1,3 +1,4 @@
+from collections import namedtuple
 from pathlib import Path as P
 from typing import Iterable, Optional
 
@@ -8,6 +9,9 @@ from git import Repo, BadName, GitCommandError
 from url_normalize.tools import deconstruct_url
 
 from commodore.git import RefError
+
+
+CommitInfo = namedtuple("CommitInfo", ["commit", "branch", "tag"])
 
 
 class Component:
@@ -113,31 +117,55 @@ class Component:
     def parameters_key(self):
         return component_parameters_key(self.name)
 
-    def checkout(self):
-        remote_heads = self._repo.remote().fetch(prune=True)
-        remote_prefix = self._repo.remote().name + "/"
-        version = self._version
-        if self._version is None:
-            # Handle case where we want the default branch of the remote
-            try:
-                version = self._repo.remote().refs["HEAD"].reference.name
-            except IndexError:
-                self._repo.git.remote("set-head", "origin", "--auto")
-                version = self._repo.remote().refs["HEAD"].reference.name
+    def _remote_prefix(self):
+        """
+        Find prefix of Git remote, will usually be 'origin/'.
+        """
+        return self._repo.remote().name + "/"
 
-            version = version.replace(remote_prefix, "", 1)
+    def _default_version(self):
+        """
+        Find default branch of the remote
+        """
+        try:
+            version = self._repo.remote().refs["HEAD"].reference.name
+        except IndexError:
+            self._repo.git.remote("set-head", "origin", "--auto")
+            version = self._repo.remote().refs["HEAD"].reference.name
+        return version.replace(self._remote_prefix(), "", 1)
+
+    def _find_commit_for_version(self, version, remote_heads):
+        remote_prefix = self._remote_prefix()
         for head in remote_heads:
-            branch = head.name
-            if branch.startswith(remote_prefix):
-                branch = branch.replace(remote_prefix, "", 1)
-            if branch == version:
+            tag = None
+            branch = None
+            headname = head.name
+            if headname.startswith(remote_prefix):
+                branch = headname.replace(remote_prefix, "", 1)
+                headname = branch
+            else:
+                tag = headname
+
+            if headname == version:
                 commit = head.commit
                 break
         else:
-            # If we haven't found a branch matching the requested version,
+            # If we haven't found a branch or tag matching the requested version,
             # assume the version is a commit sha.
             commit = version
             branch = None
+            tag = None
+
+        return CommitInfo(commit=commit, branch=branch, tag=tag)
+
+    def checkout(self):
+        remote_heads = self._repo.remote().fetch(prune=True, tags=True)
+        version = self._version
+        if self._version is None:
+            # Handle case where we want the default branch of the remote
+            version = self._default_version()
+
+        commit, branch, tag = self._find_commit_for_version(version, remote_heads)
 
         try:
             if branch:
@@ -152,6 +180,9 @@ class Component:
 
                 head.set_tracking_branch(self.repo.remote().refs[branch])
                 self._repo.head.reference = head
+            elif tag:
+                # Simply create detached head pointing to tag
+                self._repo.head.reference = tag
             else:
                 # Create detached head by setting repo.head.reference as
                 # direct ref to commit object.