+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v0.5.0] - 2021-03-12
+
 ### Added
 
 * Generate a best-effort SSH-based push URL for component repos ([#281])
@@ -245,7 +247,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial implementation
 
-[Unreleased]: https://github.com/projectsyn/commodore/compare/v0.4.2...HEAD
+[Unreleased]: https://github.com/projectsyn/commodore/compare/v0.5.0...HEAD
 [v0.1.0]: https://github.com/projectsyn/commodore/releases/tag/v0.1.0
 [v0.1.1]: https://github.com/projectsyn/commodore/releases/tag/v0.1.1
 [v0.1.2]: https://github.com/projectsyn/commodore/releases/tag/v0.1.2
@@ -261,6 +263,7 @@ Initial implementation
 [v0.4.0]: https://github.com/projectsyn/commodore/releases/tag/v0.4.0
 [v0.4.1]: https://github.com/projectsyn/commodore/releases/tag/v0.4.1
 [v0.4.2]: https://github.com/projectsyn/commodore/releases/tag/v0.4.2
+[v0.5.0]: https://github.com/projectsyn/commodore/releases/tag/v0.5.0
 
 [#53]: https://github.com/projectsyn/commodore/pull/53
 [#58]: https://github.com/projectsyn/commodore/pull/58